feat: add support for deploying to vercel.
This commit adds support for deploying to vercel.
I have changed the example project to `commonjs`, since that's easier to
support on vercel.
I have also removed the dependency on axios when running on vercel, as the package wasn't being
imported, but I can revert that change if necessary.
I have split index.js and dev_server.js as we don't want to run vite
when deploying to vercel.
I have added a flake.nix and a flake.lock, if useful I can keep them,
otherwise I am happy to remove them.