chore: return client token on production
This commit returns client token on production instead of the access
token.

We probably want a safer strategy, we should return client token unless
explicitly set that we should return the access token, if we want to use
access tokens at all.
i.e instead of NODE_ENV == production, we should return by default the
cient token, unless say something like RETURN_ACCESS_TOKEN == true. That
way if say the app is deployed publicly, but with NODE_ENV set to a
different environment (staging), we wuold not be be leaking the access
token.